make jupDialog component for consistent dialogs across the app
